fix: sync between sql and entity

+++ src/main/kotlin/com/multicampus/gamesungcoding/a11ymarketserver/feature/seller/dto/SellerDashboardResponse.kt
@@ -7,7 +7,7 @@ data class SellerDashboardResponse(
     val sellerId: UUID,
     val sellerName: String,
     val sellerIntro: String?,
-    val totalRevenue: BigDecimal?,
+    val totalRevenue: Int?,
     val totalOrderCount: Long?,
     val refundRate: BigDecimal?,
     val confirmedRate: BigDecimal?
